9095-2

+++ 소스.cpp
@@ -9,7 +9,7 @@ int main() {
     vector<int> T;
     cin >> count;
 
-    for (size_t i = 0; i < count; i++) {
+    for (int i = 0; i < count; i++) {
         int num;
         cin >> num;
         T.push_back(num);
@@ -18,7 +18,7 @@ int main() {
     memo[1] = 1;
     memo[2] = 2;
     memo[3] = 4;
-    for (size_t i = 4; i < 11; i++)
+    for (int i = 4; i < 11; i++)
         memo[i] = memo[i - 1] + memo[i - 2] + memo[i - 3];
 
     for (size_t i = 0; i < count; i++) {
